% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_lm.r
\name{arfima.sim}
\alias{arfima.sim}
\title{Simulate a time series from ARFIMA model}
\usage{
arfima.sim(mod, n)
}
\arguments{
\item{mod}{ARFIMA params (ar, d, ma, mu, sigma) after change-point}

\item{n}{length of time series}
}
\value{
a simulated time series from ARFIMA
}
\description{
Simulate a time series from ARFIMA model
}
