% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_cp.r
\name{arma.cp.sim}
\alias{arma.cp.sim}
\title{Simulate a time series from ARMA 1-change-point model}
\usage{
arma.cp.sim(mod.1, mod.2, burn.out.n = 100)
}
\arguments{
\item{mod.1}{ARMA params (ar, ma, mu, sd, n) before change-point}

\item{mod.2}{ARMA params (ar, ma, mu, sd, n) after change-point}

\item{burn.out.n}{(optional) number of samples to discard before generating the actual data}
}
\value{
a simulated time series from ARMA 1-change-point model
}
\description{
Simulate a time series from ARMA 1-change-point model
}
