% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_z2r.R
\name{convert_z2r}
\alias{convert_z2r}
\title{Convert Fisher's z into correlation coefficient r}
\usage{
convert_z2r(z)
}
\arguments{
\item{z}{Fisher's \code{z}-value.}
}
\value{
The back-transformed correlation coefficient \code{r}.
}
\description{
Convert Fisher's z into correlation coefficient r.
}
\examples{
convert_z2r(.03)

}
\references{
Lipsey MW, Wilson DB. 2001. Practical meta-analysis. Thousand Oaks, Calif: Sage Publications
            \cr \cr
            Wilson DB. 2016. Formulas Used by the "Practical Meta-Analysis Effect Size Calculator". Unpublished manuscript: George Mason University
}
