\encoding{UTF-8}
\name{esc-package}
\alias{esc-package}
\alias{esc}
\docType{package}
\title{Effect Size Computation for Meta Analysis}
\description{
  This is an R implementation of the web-based 'Practical Meta-Analysis Effect Size Calculator' from David B. Wilson.
}
\details{
\tabular{rl}{
Package: \tab esc\cr
Type: \tab Package\cr
Version: \tab 0.1.0\cr
Date: \tab 2016-06-30\cr
License: \tab GPL-3\cr
}
}
\author{
Daniel Lüdecke \email{d.luedecke@uke.de}
}
\references{
  \itemize{
    \item Github: \url{https://github.com/sjPlot/esc}
    \item Bug reports: \url{https://github.com/sjPlot/esc/issues}
  }
}
