% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esc_beta.R
\name{esc_beta}
\alias{esc_beta}
\title{Compute effect size from Standardized Regression Coefficient}
\usage{
esc_beta(beta, sdy, grp1n, grp2n, es.type = c("d", "g", "or", "logit", "r",
  "cox.or", "cox.log"))
}
\arguments{
\item{beta}{The standardized beta coefficient.}

\item{sdy}{The standard deviation of the dependent variable.}

\item{grp1n}{Treatment group sample size.}

\item{grp2n}{Control group sample size.}

\item{es.type}{Type of effect size that should be returned.
\describe{
  \item{\code{"d"}}{returns standardized mean difference effect size \code{d}}
  \item{\code{"g"}}{returns adjusted standardized mean difference effect size Hedge's \code{g}}
  \item{\code{"or"}}{returns effect size as odds ratio}
  \item{\code{"cox.or"}}{returns effect size as Cox-odds ratio (see \code{\link{esc_d2or}} for details)}
  \item{\code{"logit"}}{returns effect size as log odds}
  \item{\code{"cox.log"}}{returns effect size as Cox-log odds (see \code{\link{esc_d2logit}} for details)}
  \item{\code{"r"}}{returns correlation effect size \code{r}}
}}
}
\value{
The effect size \code{es}, the standard error \code{se}, the variance
        of the effect size \code{var}, the lower and upper confidence limits
        \code{ci.lo} and \code{ci.hi}, the weight factor \code{w} and the
        total sample size \code{totaln}.
}
\description{
Compute effect size from Standardized Regression Coefficient.
}
\note{
If \code{es.type = "r"}, Fisher's transformation for the effect size
      \code{r} and their confidence intervals are also returned.
}
\examples{
esc_beta(1.25, 3, 100, 150)
esc_beta(1.25, 3, 100, 150, es.type = "cox.log")

}
\references{
Lipsey MW, Wilson DB. 2001. Practical meta-analysis. Thousand Oaks, Calif: Sage Publications
}

