% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esc_d2or.R
\name{esc_d2or}
\alias{esc_d2or}
\title{Compute OR from effect size d}
\usage{
esc_d2or(d, se, v, info = NULL)
}
\arguments{
\item{d}{The effect size \code{d}.}

\item{se}{The standard error of \code{d}. One of \code{se} or \code{v}
must be specified.}

\item{v}{The variance of \code{d}. One of \code{se} or \code{v} must be
specified.}

\item{info}{String with information on the transformation. Used for the
print-method. Usually, this argument can be ignored}
}
\value{
The effect size \code{es}, the standard error \code{se}, the variance
        of the effect size \code{var}, the lower
        and upper confidence limits \code{ci.lo} and \code{ci.hi} as well as
        the weight factor \code{w}.
}
\description{
Compute effect size \code{OR} from effect size \code{d}.
}
\note{
Effect size is returned as \code{exp(log_values)} (odds ratio), i.e. variance,
      standard error and confidence intervals are also exponentiated. To
      get the log-values, use \code{\link{esc_d2logit}}.
}
\examples{
esc_d2or(0.7, se = 0.5)

}
\references{
Lipsey MW, Wilson DB. 2001. Practical meta-analysis. Thousand Oaks, Calif: Sage Publications
}

