% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esc_d2or.R
\name{esc_d2r}
\alias{esc_d2r}
\title{Compute correlation from effect size d}
\usage{
esc_d2r(d, se, v, grp1n, grp2n, info = NULL)
}
\arguments{
\item{d}{The effect size \code{d}.}

\item{se}{The standard error of \code{d}. One of \code{se} or \code{v}
must be specified.}

\item{v}{The variance of \code{d}. One of \code{se} or \code{v} must be
specified.}

\item{grp1n}{Treatment group sample size.}

\item{grp2n}{Control group sample size.}

\item{info}{String with information on the transformation. Used for the
print-method. Usually, this argument can be ignored}
}
\value{
The effect size \code{es}, the standard error \code{se}, the variance
        of the effect size \code{var}, the lower
        and upper confidence limits \code{ci.lo} and \code{ci.hi} as well as
        the weight factor \code{w}.
}
\description{
Compute effect size correlation from effect size \code{d}.
}
\examples{
esc_d2r(d = 0.7, se = 0.5, grp1n = 70, grp2n = 80)

}
\references{
Lipsey MW, Wilson DB. 2001. Practical meta-analysis. Thousand Oaks, Calif: Sage Publications
}

