% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esc_mean_sd.R
\name{esc_mean_se}
\alias{esc_mean_se}
\title{Compute effect size from Mean and Standard Error}
\usage{
esc_mean_se(grp1m, grp1se, grp1n, grp2m, grp2se, grp2n, es.type = c("d", "OR",
  "logit", "r"))
}
\arguments{
\item{grp1m}{The mean of the first group.}

\item{grp1se}{The standard error of the first group.}

\item{grp1n}{The sample size of the first group.}

\item{grp2m}{The mean of the second group.}

\item{grp2se}{The standard error of the second group.}

\item{grp2n}{The sample size of the second group.}

\item{es.type}{Type of effect size that should be returned. By default,
this is \code{"d"}, i.e. effect size \code{d} is returned.
Use \code{es.type = "OR"} to return effect size as odds ratios or
\code{es.type = "logit"} to return effect size as log odds,}
}
\value{
The effect size \code{es}, the standard error \code{se}, the variance
        of the effect size \code{var}, the lower
        and upper confidence limits \code{ci.lo} and \code{ci.hi} as well as
        the weight factor \code{w}.
}
\description{
Compute effect size from Mean and Standard Error.
}
\note{
If \code{es.type = "r"}, Fisher's transformation for the effect size
      \code{r} and their confidence intervals are also returned.
}
\examples{
esc_mean_se(grp1m = 7, grp1se = 1.5, grp1n = 50,
            grp2m = 9, grp2se = 1.8, grp2n = 60, es.type = "OR")

}
\references{
Lipsey MW, Wilson DB. 2001. Practical meta-analysis. Thousand Oaks, Calif: Sage Publications
}

