% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esc_or2d.R
\name{esc_or2d}
\alias{esc_or2d}
\title{Convert effect size OR from d}
\usage{
esc_or2d(or, se, v, es.type = c("d", "cox.d"), info = NULL)
}
\arguments{
\item{or}{The effect size as odds ratio.}

\item{se}{The standard error of \code{d}. One of \code{se} or \code{v}
must be specified.}

\item{v}{The variance of \code{d}. One of \code{se} or \code{v} must be
specified.}

\item{es.type}{Type of effect size that should be returned.
\describe{
  \item{\code{"d"}}{returns effect size \code{d}}
  \item{\code{"cox.d"}}{returns effect \code{d}, based on Cox method}
}}

\item{info}{String with information on the transformation. Used for the
print-method. Usually, this argument can be ignored}
}
\value{
The effect size \code{es}, the standard error \code{se}, the variance
        of the effect size \code{var}, the lower
        and upper confidence limits \code{ci.lo} and \code{ci.hi} as well as
        the weight factor \code{w}.
}
\description{
Compute effect size \code{d} from effect size \code{OR}.
}
\note{
While \code{or} is the exponentiated log odds, the variance or standard
      error need to be on the log-scale!
}
\examples{
esc_or2d(3.56, se = 0.91)
esc_d2or(0.7, se = 0.5)

}
\references{
Lipsey MW, Wilson DB. 2001. Practical meta-analysis. Thousand Oaks, Calif: Sage Publications
}

