% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esc_z2r.R
\name{esc_r2z}
\alias{esc_r2z}
\title{Convert correlation coefficient r into Fisher's z}
\usage{
esc_r2z(r)
}
\arguments{
\item{r}{The correlation coefficient.}
}
\value{
The transformed Fisher's \code{z}.
}
\description{
Convert correlation coefficient r into Fisher's z.
}
\examples{
esc_r2z(.03)

}
\references{
Lipsey MW, Wilson DB. 2001. Practical meta-analysis. Thousand Oaks, Calif: Sage Publications
            \cr \cr
            Wilson DB. 2016. Formulas Used by the "Practical Meta-Analysis Effect Size Calculator". Unpublished manuscript: George Mason University
}

