% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic.r
\name{get.bic}
\alias{get.bic}
\title{Calculate the bic value}
\usage{
get.bic(mod, n, cp.penalty)
}
\arguments{
\item{mod}{fitted model (either short memory change-point or long memory)}

\item{n}{time series length}

\item{cp.penalty}{penalty (multiplier) on change-points}
}
\value{
bic value
}
\description{
Calculate the bic value
}
\examples{
\donttest{
x1 <- arma.cp.sim(mod.1 = list(ar = 0.8, ma = 0.4, mu = 0, sd = 1, n = 100),
mod.2 = list(ar = 0.8, ma = 0.4, mu = 1, sd = 1, n = 100))
x1.es <- es(x1)
bic.cp = get.bic(x1.es$mod$cp, n = 100, cp.penalty = 1)
bic.lm = get.bic(x1.es$mod$lm, n = 100, cp.penalty = 1)
print(c(bic.cp, bic.lm))
}
}
