% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setting.r
\name{get.settings}
\alias{get.settings}
\title{Get simulation settings}
\usage{
get.settings(file.name, n)
}
\arguments{
\item{file.name}{location of the setting file}

\item{n}{simulated time series length}
}
\value{
a list of settings with corresponding sim function
}
\description{
Get simulation settings
}
