% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.r
\name{save.sim.for.resnet}
\alias{save.sim.for.resnet}
\title{Save simulated data}
\usage{
save.sim.for.resnet(setting.file, n = 512, its = 1000, root.dir)
}
\arguments{
\item{setting.file}{file for get.settings()}

\item{n}{time series length}

\item{its}{number of simulated training data for each class}

\item{root.dir}{folder for the simulated data}
}
\value{
none
}
\description{
Save simulated data
}
\examples{
\donttest{
n = 512
root.dir = "/Users/loktingyuen/Documents/sim_result/ts_classification_through_simulation/NN/512"
setting.file = "/Users/loktingyuen/Library/Mobile Documents/com~apple~CloudDocs/PhD/ESC/setting.csv"
#save.sim.for.resnet(setting.file = setting.file, n = n, its = 1000, root.dir = root.dir)
}
}
