% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.r
\name{save.sim}
\alias{save.sim}
\title{Save simulated data}
\usage{
save.sim(setting, n, its, data_folder, seed)
}
\arguments{
\item{setting}{simulation setting from get.settings()}

\item{n}{time series length}

\item{its}{number of simulated training data for each class}

\item{data_folder}{folder for the simulated data}

\item{seed}{for set.seed()}
}
\value{
none
}
\description{
Save simulated data
}
