% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esc.R
\name{effect_sizes}
\alias{effect_sizes}
\title{Generate effect size data frame from other data}
\usage{
effect_sizes(..., data, type, es.type = c("d", "g", "or", "logit", "r",
  "cox.or", "cox.log"))
}
\arguments{
\item{...}{Named arguments. The name (left-hand side) is the name of one of
\pkg{esc} functions' argument, the argument (right-hand side) is the
name of the column in \code{data} that holds the data values.
See 'Examples'.}

\item{data}{A data frame with columns that contain the values that are passed
to one of the \pkg{esc}-functions.}

\item{type}{Name of one of the \pkg{esc}-functions, where arguments in \code{...}
are passed to.}

\item{es.type}{Type of effect size that should be returned.
\describe{
  \item{\code{"d"}}{returns standardized mean difference effect size \code{d}}
  \item{\code{"g"}}{returns adjusted standardized mean difference effect size Hedges' \code{g}}
  \item{\code{"or"}}{returns effect size as odds ratio}
  \item{\code{"cox.or"}}{returns effect size as Cox-odds ratio (see \code{\link{esc_d2or}} for details)}
  \item{\code{"logit"}}{returns effect size as log odds}
  \item{\code{"cox.log"}}{returns effect size as Cox-log odds (see \code{\link{esc_d2logit}} for details)}
  \item{\code{"r"}}{returns correlation effect size \code{r}}
}}
}
\value{
A data frame with the results of multiple effect size function calls.
}
\description{
This method computes any effect size from raw values from a data
             frame. Convenient method to compute multiple effect sizes at once.
}
\note{
This function rowwise iterates \code{data} and calls the function named
        in \code{type} for the values taken from each row of \code{data}. With
        this, it is possible to compute multiple effect size results with
        just one function call, if the necessary values are stored in a data
        frame.
}
\examples{
tmp <- data.frame(
  tvalue = c(3.3, 2.9, 2.3),
  n = c(250, 200, 210),
  studyname = c("Study 1", "Study 2", "Study 3")
)

effect_sizes(t = tvalue, totaln = n, study = studyname, type = "esc_t", data = tmp)


tmp <- data.frame(
  coefficient = c(0.4, 0.2, 0.6),
  se = c(.15, .1, .2),
  treat = c(50, 60, 50),
  cntrl = c(45, 70, 40),
  author = c("Smith 2000", "Smith 2010 2", "Smith 2012")
)

effect_sizes(beta = coefficient, sdy = se, grp1n = treat, grp2n = cntrl,
    study = author, type = "esc_beta", data = tmp, es.type = "or")

}
