% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esc_beta.R
\name{esc_beta}
\alias{esc_beta}
\title{Compute effect size from Standardized Regression Coefficient}
\usage{
esc_beta(beta, sdy, grp1n, grp2n, es.type = c("d", "OR", "logit", "r"))
}
\arguments{
\item{beta}{The standardized beta coefficient.}

\item{sdy}{The standard deviation of the dependent variable.}

\item{grp1n}{Treatment group sample size.}

\item{grp2n}{Control group sample size.}

\item{es.type}{Type of effect size that should be returned. By default,
this is \code{"d"}, i.e. effect size \code{d} is returned.
Use \code{es.type = "OR"} to return effect size as odds ratios,
\code{es.type = "logit"} to return effect size as log odds and
\code{es.type = "r"} to return effect size as correlation.}
}
\value{
The effect size \code{es}, the standard error \code{se}, the variance
        of the effect size \code{var}, the lower
        and upper confidence limits \code{ci.lo} and \code{ci.hi} as well as
        the weight factor \code{w}.
}
\description{
Compute effect size from Standardized Regression Coefficient.
}
\examples{
esc_beta(1.25, 3, 100, 150)
esc_beta(1.25, 3, 100, 150, es.type = "logit")

}
\references{
Lipsey MW, Wilson DB. 2001. Practical meta-analysis. Thousand Oaks, Calif: Sage Publications
}

