% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esc_mean_sd.R
\name{esc_mean_sd}
\alias{esc_mean_sd}
\title{Compute effect size from Mean and Standard Deviation}
\usage{
esc_mean_sd(grp1m, grp1sd, grp1n, grp2m, grp2sd, grp2n, es.type = c("d", "g",
  "or", "logit", "r", "cox.or", "cox.log"))
}
\arguments{
\item{grp1m}{The mean of the first group.}

\item{grp1sd}{The standard deviation of the first group.}

\item{grp1n}{The sample size of the first group.}

\item{grp2m}{The mean of the second group.}

\item{grp2sd}{The standard deviation of the second group.}

\item{grp2n}{The sample size of the second group.}

\item{es.type}{Type of effect size that should be returned.
\describe{
  \item{\code{"d"}}{returns standardized mean difference effect size \code{d}}
  \item{\code{"g"}}{returns adjusted standardized mean difference effect size Hedge's \code{g}}
  \item{\code{"or"}}{returns effect size as odds ratio}
  \item{\code{"cox.or"}}{returns effect size as Cox-odds ratio (see \code{\link{esc_d2or}} for details)}
  \item{\code{"logit"}}{returns effect size as log odds}
  \item{\code{"cox.log"}}{returns effect size as Cox-log odds (see \code{\link{esc_d2logit}} for details)}
  \item{\code{"r"}}{returns correlation effect size \code{r}}
}}
}
\value{
The effect size \code{es}, the standard error \code{se}, the variance
        of the effect size \code{var}, the lower
        and upper confidence limits \code{ci.lo} and \code{ci.hi} as well as
        the weight factor \code{w}.
}
\description{
Compute effect size from Mean and Standard Deviation.
}
\note{
If \code{es.type = "r"}, Fisher's transformation for the effect size
      \code{r} and their confidence intervals are also returned.
}
\examples{
esc_mean_sd(grp1m = 7, grp1sd = 2, grp1n = 50,
            grp2m = 9, grp2sd = 3, grp2n = 60, es.type = "logit")

}
\references{
Lipsey MW, Wilson DB. 2001. Practical meta-analysis. Thousand Oaks, Calif: Sage Publications
}

