% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esc_t.R
\name{esc_t}
\alias{esc_t}
\title{Compute effect size from Student's t-test}
\usage{
esc_t(t, p, totaln, grp1n, grp2n, es.type = c("d", "OR", "logit", "r"))
}
\arguments{
\item{t}{The t-value of the t-test. One of \code{t} or \code{p} must be specified.}

\item{p}{The p-value of the t-test. One of \code{t} or \code{p} must be specified.}

\item{totaln}{Total sample size. Either \code{totaln}, or \code{grp1n} and
\code{grp2n} must be specified.}

\item{grp1n}{Treatment group sample size.}

\item{grp2n}{Control group sample size.}

\item{es.type}{Type of effect size that should be returned. By default,
this is \code{"d"}, i.e. effect size \code{d} is returned.
Use \code{es.type = "OR"} to return effect size as odds ratios,
\code{es.type = "logit"} to return effect size as log odds and
\code{es.type = "r"} to return effect size as correlation.}
}
\value{
The effect size \code{es}, the standard error \code{se}, the variance
        of the effect size \code{var}, the lower
        and upper confidence limits \code{ci.lo} and \code{ci.hi} as well as
        the weight factor \code{w}.
}
\description{
Compute effect size from Student's t-test for \emph{independent samples}.
}
\note{
This function only applies to \emph{independent sample} t-tests, either
      equal or unequal sample sizes. It can't be used for t-values from
      dependent or paired t-tests, or t-values from other statistical procedures
      (like regressions).
}
\examples{
# unequal sample size
esc_t(t = 3.3, grp1n = 100, grp2n = 150)

# equal sample size
esc_t(t = 3.3, totaln = 200)

# unequal sample size, with p-value
esc_t(p = 0.03, grp1n = 100, grp2n = 150)

# equal sample size, with p-value
esc_t(p = 0.03, totaln = 200)

}
\references{
Lipsey MW, Wilson DB. 2001. Practical meta-analysis. Thousand Oaks, Calif: Sage Publications
}

