% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hedges_g.R
\name{hedges_g}
\alias{hedges_g}
\title{Compute Hedges' g from effect size d}
\usage{
hedges_g(es, totaln)
}
\arguments{
\item{es}{The effect size \code{d}.}

\item{totaln}{The total sample size.}
}
\value{
The Hedges' g coefficient.
}
\description{
Compute Hedges' g from effect size d.
}
\examples{
hedges_g(0.75, 50)

}
\references{
Lipsey MW, Wilson DB. 2001. Practical meta-analysis. Thousand Oaks, Calif: Sage Publications
}

